pyproject.toml
[project]
name = "hike"
version = "1.5.0"
description = "A Markdown browser for the terminal"
authors = [
    { name = "Dave Pearson", email = "davep@davep.org" },
    { name = "Tobias Hochguertel" },
]
maintainers = [
    { name = "Tobias Hochguertel" },
]
dependencies = [
    "textual>=5.3.0",
    "textual-enhanced>=0.13.0",
    "textual-fspicker>=0.4.1",
    "xdg-base-dirs>=6.0.2",
    "httpx>=0.28.1",
    "pathspec>=0.12.1",
    "typing-extensions>=4.12.2",
    "pyperclip>=1.9.0",
    "typer>=0.24.1",
    "pydantic>=2.13.2",
    "pydantic-settings>=2.13.1",
    "ruamel-yaml>=0.19.1",
    "python-dotenv>=1.2.2",
]
readme = "README.md"
requires-python = ">= 3.10"
license = "GPL-3.0-or-later"
license-files = ["LICENSE"]
keywords = [
    "terminal",
    "tui",
    "chui",
    "documentation",
    "markdown",
    "markdown-viewer",
    "file-viewer",
    "file-browser",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Other Audience",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Documentation",
    "Topic :: Terminals",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://tobiashochguertel.github.io/hike/"
Repository = "https://github.com/tobiashochguertel/hike"
Documentation = "https://tobiashochguertel.github.io/hike/"
Source = "https://github.com/tobiashochguertel/hike"
Issues = "https://github.com/tobiashochguertel/hike/issues"
Upstream = "https://github.com/davep/hike"

[project.scripts]
hike = "hike.__main__:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.uv]
managed = true

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pre-commit>=4.1.0",
    "textual-dev>=1.7.0",
    "pytest>=8.3.4",
    "mkdocs-material>=9.6.11",
    "markdown-exec>=1.10.3",
    "pygments-ansi-color>=0.3.0",
    "codespell>=2.4.1",
    "ruff>=0.12.9",
    "libtmux>=0.55.0",
    "hypothesis>=6.152.1",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.pyright]
venvPath="."
venv=".venv"
exclude=[".venv"]

[tool.pytest.ini_options]
markers = [
    "tmux: end-to-end tests that drive Hike inside an isolated tmux server",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    # I think try...expect...pass reads far better.
    "SIM105",
]

[tool.ruff.lint.per-file-ignores]
"src/hike/cli/*.py" = ["B008"]

[tool.ruff.lint.pycodestyle]
max-line-length = 120
